rename VideoPress video folder
