+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/poster-image-block-control/index.js
