+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/videopress-uploader/uploader-editor.js
