+++ projects/packages/videopress/src/client/block-editor/blocks/video/hooks/use-poster-image.js
