+++ projects/packages/videopress/src/client/block-editor/blocks/video/scripts/vp-block-bridge/index.js
