+++ projects/packages/videopress/src/client/block-editor/index.js
@@ -1,4 +1,4 @@
 /**
  * Internal dependencies
  */
-import './blocks/videopress/';
+import './blocks/video';
